Evaluate expression before initializing variable
Fixes a bug where the program would segfault if the same
variable that was being initialized was used in the expression,
since the symbol had been created but no value had yet been assigned.

The fix was to move the evaluation of the expression before
the creation of the symbol.